Renamed downloadAPK to downloader
